pyproject.toml
[project]
name = "mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=6.30.0",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
]
